+++ lib/ui/widgets/signature_type_picker.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:widgetbook/widgetbook.dart';
 import 'package:widgetbook_annotation/widgetbook_annotation.dart' as widgetbook;
 
 import '../../data/signature_type.dart';
@@ -6,17 +7,22 @@ import '../../strings_context.dart';
 import '../app_theme.dart';
 import 'certificate_picker.dart';
 
-/// Displays two options to select the [SignatureType] options.
+/// Displays two options to select the [SignatureType] options; either
+/// [SignatureType.withTimestamp] or [SignatureType.withoutTimestamp].
+///
+/// When [canChange] is `false`, value selection by user is disabled.
 ///
 /// See also:
 ///  - [CertificatePicker]
 class SignatureTypePicker extends StatelessWidget {
   final SignatureType? value;
+  final bool canChange;
   final ValueChanged<SignatureType> onValueChanged;
 
   const SignatureTypePicker({
     super.key,
     required this.value,
+    this.canChange = true,
     required this.onValueChanged,
   });
 
@@ -36,6 +42,7 @@ class SignatureTypePicker extends StatelessWidget {
   Widget _listItem(SignatureType value) {
     return _ListItem(
       value: value, // value from param
+      canSelect: canChange,
       selectedValue: this.value, // value from Widget
       onSelected: () {
         onValueChanged(value);
@@ -47,11 +54,13 @@ class SignatureTypePicker extends StatelessWidget {
 /// [SignatureTypePicker] - [ListView] item.
 class _ListItem extends StatelessWidget {
   final SignatureType value;
+  final bool canSelect;
   final SignatureType? selectedValue;
   final VoidCallback onSelected;
 
   const _ListItem({
     required this.value,
+    required this.canSelect,
     required this.selectedValue,
     required this.onSelected,
   });
@@ -62,24 +71,29 @@ class _ListItem extends StatelessWidget {
     final titleText = strings.signatureTypeValueTitle(value.name);
     final subtitleText = strings.signatureTypeValueSubtitle(value.name);
 
-    // NOT using RadioListTile because need to scale-up and style Radio
-
-    return ListTile(
-      onTap: onSelected,
-      leading: Transform.scale(
-        scale: kRadioScale,
-        child: Radio<SignatureType>(
-          value: value,
-          groupValue: selectedValue,
-          onChanged: (final SignatureType? value) {
-            if (value != null) {
-              if (selectedValue != value) {
-                onSelected();
+    final enabled = (canSelect || value == selectedValue);
+    final radio = Radio<SignatureType>(
+      value: value,
+      groupValue: selectedValue,
+      onChanged: enabled
+          ? (final SignatureType? value) {
+              if (value != null) {
+                if (selectedValue != value) {
+                  onSelected();
+                }
               }
             }
-          },
-          activeColor: kRadioActiveColor,
-        ),
+          : null,
+      activeColor: kRadioActiveColor,
+    );
+
+    // NOT using RadioListTile because need to scale-up and style Radio
+    return ListTile(
+      onTap: (canSelect ? onSelected : null),
+      enabled: enabled,
+      leading: Transform.scale(
+        scale: kRadioScale,
+        child: radio,
       ),
       title: Text(
         titleText,
@@ -92,16 +106,24 @@ class _ListItem extends StatelessWidget {
 
 @widgetbook.UseCase(
   path: '[Lists]',
-  name: 'SignatureTypePicker',
+  name: '',
   type: SignatureTypePicker,
 )
 Widget previewSignatureTypePicker(BuildContext context) {
-  SignatureType? selectedValue;
+  final bool canChange = context.knobs.boolean(
+    label: "Can change",
+    initialValue: true,
+  );
+  SignatureType? selectedValue = context.knobs.listOrNull(
+    label: "Signature type",
+    options: [SignatureType.withTimestamp, SignatureType.withoutTimestamp],
+  );
 
   return StatefulBuilder(
     builder: (context, setState) {
       return SignatureTypePicker(
         value: selectedValue,
+        canChange: canChange,
         onValueChanged: (value) {
           setState(() => selectedValue = value);
         },